fixed debug mode

+++ libs/ma/inc/module/sweepSvJumps.h
@@ -650,8 +650,7 @@ class FilterDiagonalLineCalls
 {
   public:
     int64_t iFilterDiagonalLineCalls;
-    FilterDiagonalLineCalls( const ParameterSetManager& rParameters )
-        : iFilterDiagonalLineCalls( 300 )
+    FilterDiagonalLineCalls( const ParameterSetManager& rParameters ) : iFilterDiagonalLineCalls( 300 )
     {} // constructor
 
     inline int64_t getStd( std::vector<int64_t>& vX )
@@ -685,7 +684,7 @@ class FilterDiagonalLineCalls
             } // for
             int64_t iStdA, iStdB;
             iStdA = getStd( vDiagonalA );
-            iStdB = std::max(getStd( vDiagonalB ), 1);
+            iStdB = std::max( getStd( vDiagonalB ), (int64_t)1 );
             if( iStdA / iStdB < iFilterDiagonalLineCalls )
                 pRet->vContent.push_back( pCall );
         } // for
